Add pyright and setting (#56)
* chore: add pyright for dev

* chore: configure pyright

* feat: add pyright and nodeenv packages to dependencies

* chore: update VSCode settings to enable pyright

* feat: setup and run pyright by github ci

* style: rename job name to ci-checks

* fix: update audio duration constants to use float values

* fix: change comp_duration type from int to float for better precision

* fix: change timeline type to SpeechTimeline

* fix: typecast np.argmax(correlation) to int

* fix: change create_srt_files return type to None

* fix: change start_offset type from int to float for better precision

* chore: add Node.js setup step in CI workflow

* fix: use pathlib.path on clip_test

* fix: use pathlib.path on audio_files_test

* fix: use pathlib.path on segments_test

* fix: use pathlib.path on srt_test

* chore: remove Pyright installation from CI workflow

+++ src/waddle/audios/align_offset.py
@@ -20,7 +20,7 @@ def find_offset_via_cross_correlation(ref_audio: np.ndarray, spk_audio: np.ndarr
     Negative offset => spk_audio starts later than ref_audio
     """
     correlation = signal.correlate(ref_audio, spk_audio, mode="full")
-    max_corr_index = np.argmax(correlation)
+    max_corr_index = int(np.argmax(correlation))
     offset = max_corr_index - (
         len(spk_audio) - 1
     )  # the center index (zero lag) is at len(spk_audio) - 1
@@ -58,7 +58,7 @@ def align_speaker_to_reference(
     speaker_path: Path,
     output_dir: Path = Path("out"),
     sample_rate: int = DEFAULT_SR,
-    comp_duration: int = DEFAULT_COMP_AUDIO_DURATION,
+    comp_duration: float = DEFAULT_COMP_AUDIO_DURATION,
 ) -> Path:
     # 1) Load short segments for cross-correlation
     ref_audio, _ = librosa.load(
